@@ -35,7 +35,7 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": "^8.0",
+        "php": "^8.2",
         "dompdf/dompdf": "^3.1",
         "guzzlehttp/guzzle": "^7.10",
         "psr/container": "^1.0",
